requirements.txt
# FastAPI and Server
fastapi==0.111.0
uvicorn[standard]==0.29.0
# HTTP Client
httpx==0.27.0
# Database
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
# Pydantic (FastAPI dependency)
pydantic==2.7.4
# Testing (Optional but recommended)
pytest==8.2.1
pytest-mock==3.12.0
